fix: goalId 경로를 통해 주입(#48)

+++ src/main/java/woojooin/planit/domain/goal/mapper/GoalMapper.java
@@ -22,6 +22,6 @@ public interface GoalMapper {
 
 	int deleteGoal(@Param("objectId") Long objectId, @Param("memberId") Long memberId);
 
-	List<GoalProgressGraphDTO> selectGoalProgressByGoalId(int goalId);
+	List<GoalProgressGraphDTO> selectGoalProgressByGoalId(Long goalId);
 
 }
